fixed msvc compiler errors

+++ src/chess_engine/board.cpp
@@ -4,7 +4,7 @@
 namespace chess {
 
 Piece Board::get_piece(const Square& square) const {
-  const u_int8_t index = square.rank * BOARD_SIZE + square.file;
+  const uint8_t index = square.rank * BOARD_SIZE + square.file;
   return board[index];
 }
 void Board::set_piece(const Square& square, Piece piece) {
@@ -18,7 +18,7 @@ bool Board::is_empty(Square square) const { return get_piece(square).type == Pie
 
 std::vector<Square> Board::get_pieces(Piece::Type type, Piece::Color color) const {
   std::vector<Square> result;
-  for (size_t i = 0; i < BOARD_SIZE * BOARD_SIZE; ++i) {
+  for (uint8_t i = 0; i < BOARD_SIZE * BOARD_SIZE; ++i) {
     if (board[i].type == type && board[i].color == color) {
       result.push_back(Square(i / BOARD_SIZE, i % BOARD_SIZE));
     }
@@ -28,7 +28,7 @@ std::vector<Square> Board::get_pieces(Piece::Type type, Piece::Color color) cons
 
 std::vector<Square> Board::get_pieces(Piece::Color color) const {
   std::vector<Square> result;
-  for (size_t i = 0; i < BOARD_SIZE * BOARD_SIZE; ++i) {
+  for (uint8_t i = 0; i < BOARD_SIZE * BOARD_SIZE; ++i) {
     if (board[i].color == color) {
       result.push_back(Square(i / BOARD_SIZE, i % BOARD_SIZE));
     }
